refactor(ui): delete redundant view shims
